#288 refactor: ヘッダーの共通部分を切り出す
